+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - `Operation` includes now possible operation name which allows to support
   documents with multiple operations.
+- `Language.GraphQL.Encoder.document` and other encoding functions take a
+  `Formatter` as argument to distinguish between minified and pretty printing.
 
 ### Removed
 - Unused `Language.GraphQL.Encoder.spaced`.